pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.sdist]
exclude = [ ".conda" ]

[project]
name = "photoboxy"
version = "0.0.1"
authors = [
  { name="Chris Lee", email="github@chrislee.dhs.org" },
]
description = "A static photo album generator"
readme = "README.md"
requires-python = ">=3.11"
classifiers = [
    'Environment :: Console',
    'Intended Audience :: End Users/Desktop',
    'License :: OSI Approved :: GNU General Public License v3 or later (GPLv3+)',
    'Programming Language :: Python :: 3 :: Only',
    'Natural Language :: English',
    'Operating System :: POSIX :: Linux',
    'Topic :: Multimedia :: Graphics'
]

dependencies = [
  "pillow>=10.4",
  "jinja2>=3.1",
  "typer>=0.12",
  "numpy>=2.1",
  "insightface>=0.7.3",
  "scikit-learn>=1.5",
  "onnxruntime>=1.19",
  "diskcache>=5.6",
  "unoconv>=0.9.0",
  "gpsphoto>=2.2.3",
  "piexif>=1.1.3",
  "exifread>=3.0.0"
]

[project.scripts]
photoboxy = "photoboxy.cli:app"

[project.urls]
Homepage = "https://github.com/chrislee35/photoboxy"
Issues = "https://github.com/chrislee35/photoboxy/issues"